+++ projects/plugins/jetpack/tests/php/modules/widgets/Jetpack_Display_Posts_Widget_Test.php
@@ -147,7 +147,8 @@ class Jetpack_Display_Posts_Widget_Test extends WP_UnitTestCase {
 				'code' => 200,
 			),
 			'body'     => json_encode(
-				array( 'error' => 'test error' )
+				array( 'error' => 'test error' ),
+				JSON_UNESCAPED_SLASHES
 			),
 		);
 
@@ -170,7 +171,8 @@ class Jetpack_Display_Posts_Widget_Test extends WP_UnitTestCase {
 				'code' => 200,
 			),
 			'body'     => json_encode(
-				array( 'mydata' => 'your data' )
+				array( 'mydata' => 'your data' ),
+				JSON_UNESCAPED_SLASHES
 			),
 		);