+++ projects/js-packages/publicize-components/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2025-12-08
+### Added
+- Add the new media selection UI for Social. [#46085]
+- Add unified modal integration for Social. [#46102]
+
+### Changed
+- Social: Update empty connections UI in the editor. [#46219]
+- Update connections list in the editor sidebar to flip the toggle position. [#46218]
+- Update package dependencies. [#45914]
+
 ## [1.7.1] - 2025-12-01
 ### Added
 - Add unified UI v1 feature flag. [#46061]
@@ -1440,6 +1450,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Updated package dependencies. [#24470]
 
+[1.8.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.7.1...v1.8.0
 [1.7.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.6.1...v1.7.0
 [1.6.1]: https://github.com/Automattic/jetpack-publicize-components/compare/v1.6.0...v1.6.1